% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{countsIso}
\alias{countsIso}
\title{create count tables from isomirs}
\usage{
countsIso(x, ref = FALSE, iso5 = FALSE, iso3 = FALSE, add = FALSE,
  mism = FALSE, seed = FALSE)
}
\arguments{
  \item{x}{IsomirDataSeq class}

  \item{ref}{differenciate reference miRNA from rest}

  \item{iso5}{differenciate trimming at 5 miRNA from rest}

  \item{iso3}{differenciate trimming at 3 miRNA from rest}

  \item{add}{differenciate additions miRNA from rest}

  \item{mism}{differenciate nt substitution miRNA from
  rest}

  \item{seed}{differenciate changes in 2-7 nt from rest}
}
\value{
count table
}
\description{
create count tables from isomirs
}
\examples{
library(DESeq2)
data(isomiRex)
ma<-IsoCounts(isomiRex)
}

