% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat286.long}
\alias{dat286.long}
\title{Data frame containing mirna from Argyropoulos's paper}
\format{mirna expression data in long format.}
\usage{
dat286.long
}
\description{
Argyropoulos, Christos, et al. "Modeling bias and variation in
the stochastic processes of small RNA sequencing."
Nucleic Acids Research (2017).
}
\keyword{datasets}
