% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{design}
\alias{design}
\alias{design,IsomirDataSeq-method}
\alias{design<-,IsomirDataSeq,formula-method}
\alias{design}
\title{Accessors for the 'design' slot of a IsomirDataSeq object.}
\usage{
\S4method{design}{IsomirDataSeq}(object)

\S4method{design}{IsomirDataSeq,formula}(object) <- value
}
\arguments{
\item{object}{a \code{IsomirDataSeq} object}

\item{value}{a \code{formula} to pass to DESeq2}
}
\value{
design for the experiment
}
\description{
The design holds the R \code{formula} which expresses how the
counts depend on the variables in \code{colData}.
See \code{\link{IsomirDataSeq}} for details.
}
\examples{

data(mirData)
design(mirData) <- formula(~ 1)
}
