% Generated by roxygen2 (4.0.0): do not edit by hand
\name{do.mir.table}
\alias{do.mir.table}
\title{do counts table considering what isomiRs take into account}
\usage{
do.mir.table(x, ref = F, iso5 = F, iso3 = F, add = F, mism = F,
  seed = F)
}
\arguments{
\item{x}{object isomiDataSeq}

\item{ref}{differenciate reference miRNA from rest}

\item{iso5}{differenciate trimming at 5 miRNA from rest}

\item{iso3}{differenciate trimming at 3 miRNA from rest}

\item{add}{differenciate additions miRNA from rest}

\item{mism}{differenciate nt substitution miRNA from rest}

\item{seed}{differenciate changes in 2-7 nt from rest}
}
\description{
do counts table considering what isomiRs take into account
}

