% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ego}
\alias{ego}
\title{enrichResult class}
\format{enrichResult class with the output of:
ego <- enrichGO(row.names(assay(gene_ex_rse, "norm")),
org.Mm.eg.db, "ENSEMBL", ont = "BP")}
\usage{
ego
}
\description{
enrichResult class
}
\keyword{datasets}
