% Generated by roxygen2 (4.0.0): do not edit by hand
\name{filter.table}
\alias{filter.table}
\title{Filter tablo reference}
\usage{
filter.table(table, cov = 10)
}
\arguments{
\item{table}{object miraligner table}

\item{cov}{remove sequences that have relative abundance lower than this number}
}
\description{
Filter tablo reference
}

