% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{find_targets}
\alias{find_targets}
\title{Find miRNAs target using mRNA/miRNA expression}
\usage{
find_targets(mirna_rse, gene_rse, target, summarize = "group",
  min_cor = -0.6)
}
\arguments{
\item{mirna_rse}{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} with miRNA
information. See details.}

\item{gene_rse}{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} with gene
information. See details.}

\item{target}{Matrix with miRNAs (columns) and genes (rows)
target prediction values (1 if it is a target, 0 if not).}

\item{summarize}{Character column name in colData(rse) to use to group
samples and compare betweem miRNA/gene expression.}

\item{min_cor}{Numeric cutoff for correlation value that
will be use to consider a miRNA-gene pair as valid.}
}
\value{
mirna-gene matrix
}
\description{
This function creates a matrix with rows (genes) and
columns (mirnas) with values indicating if miRNA-gene
pair is target according putative targets and negative
correlation of the expression of both molecules.
}
\examples{

pairs <- as.matrix(data.frame(row.names=c("gene1", "gene2"),
                            mirna1=c(0,1), mirna2=c(1,0)))
mirna_matrix <- as.matrix(data.frame(row.names=c("mirna1", "mirna2"),
                                    time0_1=c(1,1),time0_2=c(1.2,0.9),
                                    time1_1=c(8,8),time1_2=c(8.2,7.9)))
gene_matrix <- as.matrix(data.frame(row.names=c("gene1", "gene2"),
                                    time0_1=c(8,8),time0_2=c(8.2,7.9),
                                    time1_1=c(1,1),time1_2=c(1.2,0.9)))
mirna_col <- data.frame(row.names=c("time0_1","time0_2","time1_1","time1_2"),
                       group=c("t0","t0","t1","t1"))
gene_col <- data.frame(row.names=c("time0_1","time0_2","time1_1","time1_2"),
                       group=c("t0","t0","t1","t1"))

mirna <- SummarizedExperiment(assays=SimpleList(norm=as.matrix(mirna_matrix)),
                            colData= mirna_col)
gene <- SummarizedExperiment(assays=SimpleList(norm=as.matrix(gene_matrix)),
                             colData= gene_col)
find_targets(mirna, gene, pairs)
}
