% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{findTargets}
\alias{findTargets}
\title{Find miRNAs target using mRNA/miRNA expression}
\usage{
findTargets(mirna_rse, gene_rse, target, summarize = "group",
  min_cor = -0.6)
}
\arguments{
\item{mirna_rse}{\code{SummarizedExperiment} with miRNA
information. See details.}

\item{gene_rse}{\code{SummarizedExperiment} with gene
information. See details.}

\item{target}{Data.frame with two columns: gene and miRNA.}

\item{summarize}{Character column name in colData(rse) to use to group
samples and compare betweem miRNA/gene expression.}

\item{min_cor}{Numeric cutoff for correlation value that
will be use to consider a miRNA-gene pair as valid.}
}
\value{
mirna-gene matrix
}
\description{
This function creates a matrix with rows (genes) and
columns (mirnas) with values indicating if miRNA-gene
pair is target according putative targets and negative
correlation of the expression of both molecules.
}
\examples{

data(isoExample)
mirna_ma <- data.frame(gene = names(gene_ex_rse)[1:20],
                       mir = names(mirna_ex_rse))
corMat <- findTargets(mirna_ex_rse, gene_ex_rse, mirna_ma)
}
