% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_ex_rse}
\alias{gene_ex_rse}
\title{Data frame containing gene expression data}
\format{gene expression data with 18 samples:
example of a time series data}
\usage{
gene_ex_rse
}
\description{
Data frame containing gene expression data
}
\keyword{datasets}
