% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoAnnotate}
\alias{isoAnnotate}
\title{Annotate the rawData of the \link{IsomirDataSeq} object}
\usage{
isoAnnotate(ids)
}
\arguments{
\item{ids}{Object of class \link{IsomirDataSeq}.}
}
\value{
\link{data.frame} with the sequence, isomir name,
and importance for each sample and isomiR.
}
\description{
Get the sequence and the name information for each isomiR,
and the importance value (\code{isomir_reads/mirna_reads}) for
each  sample.
}
\details{
\code{edit_mature_position} represents the position at the mature
sequence + nucleotide at reference + nucleotide at isomiR.
}
\examples{
data(mirData)
head(isoAnnotate(mirData))
}
