% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{isoCorrect}
\alias{isoCorrect}
\title{Calculate the correction factors from the equimolar 286 experiments}
\usage{
isoCorrect(train, data, cycles = 5000)
}
\arguments{
\item{train}{long data.frame to train model}

\item{data}{long data.frame to correct abundance}

\item{cycles}{number of cycles to reach convergency}
}
\value{
data.frame with corrected expression
}
\description{
this is the source file for fitting the linear quadratic normal famile
change the string to the location you saved this file in your drive
}
\examples{
options(warn = -1)
data(mirTritation)
ma <- isoCorrect(mirTritation[mirTritation$class=="train",],
mirTritation[mirTritation$class=="test",],cycles=5)
library(ggplot2)
ggplot(ma,aes(y=log2(reads), x=Dilution)) + geom_jitter()
ggplot(ma,aes(y=m, x=Dilution)) + geom_jitter()
}
