% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isomir.general.type}
\alias{isomir.general.type}
\title{plot general information}
\usage{
isomir.general.type(table, colid)
}
\arguments{
\item{table}{object miraligner table}

\item{colid}{column to be analize}
}
\description{
plot general information
}

