% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isomir.position}
\alias{isomir.position}
\title{Do summary of different isomiRs events}
\usage{
isomir.position(table, colid)
}
\arguments{
\item{table}{object miraligner table}

\item{colid}{column to be considered}
}
\description{
Do summary of different isomiRs events
}

