% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRexp-data.R
\docType{data}
\name{isomiRexp}
\alias{isomiRexp}
\title{IsomirDataSeq}
\format{a \code{IsomirDataSeq} class.}
\source{
Data was download from GEO dataset and analyzed with seqbuster tool.
adrec file adapter 8 1 34
miraligner file databse hsa 1 3 3 prefix
}
\usage{
data("isomiRexp")
}
\description{
This data set is the object return by \code{IsomirDataSeqFromFiles}. It mainly
contains miRNA sequencing raw data from 6 samples: 3 newnborns and 3 elderly
human individuals (Somel et al, 2010).
Use \code{colData} to see the experiment design of the data.
}
\author{
Lorena Pantano, 2015-05-19
}
\references{
Mehmet Somel et al. MicroRNA, mRNA, and protein expression link
development and aging in human and macaque brain.
Genome Research, 20(9):1207–1218, 2010.doi:10.1101/gr.106849.110.
}

