% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomiRs-package.R
\name{dLQNO}
\alias{dLQNO}
\title{Deprecated functions in package isomiRs}
\usage{
dLQNO()
}
\description{
Deprecated functions in package isomiRs
}
\details{
The following functions are deprecated and will be made defunct;
use the replacement indicated below:
\itemize{
\item isoCorrection dLQNO qLQNO pLQNO rLQNOLQNO isoLQNO
}
}
