 \name{isomiRs-internal} 
        \alias{filter.by.cov} 
        \alias{put.header} 
         
        \title{Internal mypackage Functions} 
        \description{ 
          Internal mypackage functions 
        } 
        \details{ 
          These are not to be called by the user. 
        } 
        \keyword{ internal } 
        