% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomiRs-package.R
\docType{package}
\name{isomiRs-package}
\alias{isomiRs}
\alias{isomiRs-package}
\title{isomiRs}
\description{
Characterization of miRNAs and isomiRs, clustering and
differential expression.
}
\author{
\strong{Maintainer}: Lorena Pantano \email{lorena.pantano@gmail.com}

Authors:
\itemize{
  \item Georgia Escaramis
}

}
