% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{isoNetwork}
\alias{isoNetwork}
\title{Clustering miRNAs-genes pairs in similar pattern expression}
\usage{
isoNetwork(mirna_rse, gene_rse, target, summarize = "group", org,
  genename = "ENSEMBL", min_cor = -0.6)
}
\arguments{
\item{mirna_rse}{\link[SummarizedExperiment]{SummarizedExperiment} with miRNA
information. See details.}

\item{gene_rse}{\link[SummarizedExperiment]{SummarizedExperiment} with gene
information. See details.}

\item{target}{matrix with miRNAs (columns) and genes (rows)
target prediction (1 if it is a target, 0 if not).}

\item{summarize}{character column name in colData(rse) to use to group
samples and compare betweem miRNA/gene expression.}

\item{org}{\link[AnnotationDbi]{AnnotationDb}. (org.Mm.eg.db)}

\item{genename}{character keytype of the gene
names in gene_rse object.}

\item{min_cor}{numeric cutoff to consider a miRNA to regulate a target}
}
\description{
Clustering miRNAs-genes pairs in similar pattern expression
}
\details{
This function will correlate miRNA and gene expression data using
a specific metadata variable to group samples and detect pattern
of expression that will be annotated with GO terms.
mirna_rse and gene_rse can be created using the following code:

\code{mi_rse = SummarizedExperiment(assays=SimpleList(norm=mirna_matrix), colData, metadata=list(sign=mirna_keep))}

where, \code{mirna_matrix} is the normalized counts expression,
\code{colData} is the metadata information and \code{mirna_keep}
the list of miRNAs to be used by this function.
}
\examples{

library(org.Mm.eg.db)
library(clusterProfiler)
data(isoExample)
ego <- enrichGO(row.names(assay(gene_ex_rse, "norm")), org.Mm.eg.db, ont = "BP", keytype="ENSEMBL")
data = isoNetwork(mirna_ex_rse, gene_ex_rse, ma_ex, org=ego@result)
isoPlotNet(data)
}

