% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoNorm}
\alias{isoNorm}
\title{normalize count data}
\usage{
isoNorm(x, formula = ~condition)
}
\arguments{
  \item{x}{IsomirDataSeq object}

  \item{formula}{formula that will be used for DE}
}
\description{
normalize count data
}
\examples{
library(DESeq2)
data(isomiRexp)
ma<-isoNorm(isomiRexp, formula=~condition)
}

