% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoPlotPosition}
\alias{isoPlotPosition}
\title{Plot changes at a given position}
\usage{
isoPlotPosition(ids, position = 1, column = "condition")
}
\arguments{
\item{ids}{object of class \code{\link{IsomirDataSeq}}}

\item{position}{integer indicating the position to show}

\item{column}{string indicating the column in
colData to color samples.}
}
\value{
\code{\link{IsomirDataSeq}} with new stored data to avoid
same calculation in the future.
}
\description{
This function plot different isomiRs proportion for each sample at a given
position focused on the nucleotide changes that happen there.
}
\details{
It shows the nucleotides changes at the given position for each
sample in each group.
The color indicates the sample group. The size of the point is proportional
to the number of total counts with changes.
The position at \code{y} is the number of different sequences.
}
\examples{
data(mirData)
isoPlotPosition(mirData)
}

