% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{isoPLSDA}
\alias{isoPLSDA}
\title{Partial Least Squares Discriminant Analysis for isomirSeqData}
\usage{
isoPLSDA(ids, group, validation = NULL, learn = NULL, test = NULL,
 tol = 0.001, nperm = 400, refinment = FALSE, vip = 1.2)
}
\arguments{
\item{ids}{object of class \code{\link{IsomirDataSeq}}}

\item{group}{column name in design data.frame}

\item{validation}{type of validation, either NULL or "learntest".
Default NULL}

\item{learn}{optional vector of indexes for a learn-set.
Only used when validation="learntest". Default NULL}

\item{test}{optional vector of indices for a test-set.
Only used when validation="learntest". Default NULL}

\item{tol}{tolerance value based on maximum change of cumulative R-squared
coefficient for each additional PLS component. Default tol=0.001}

\item{nperm}{number of permutations to compute the PLD-DA p-value b
ased on R2 magnitude. Default nperm=400}

\item{refinment}{logical indicating whether a refined model, based on
filtering out variables with low VIP values}

\item{vip}{Variance Importance in Projection threshold value when
a refinement precess is considered. Default vip=1.2}
}
\value{
PLS model

\code{\link[base]{list}} with the following elements: R2Matrix
(R-squared coefficients of the PLS model),
components (of the PLS),
p.value obtained by the permutations and R2PermutationVector with all R2 from
the permutations.

If the option \code{refinment} is set to \code{TRUE}, then the following
elements will appear:
R2RefinedMatrix and componentsRefinedModel (R-squared coefficients and components
 (components of the PLS model only using the most important miRNAs). As well,
p.valRefined and R2RefinedPermutationVector with the p-value and the R2 of the
permutations shuffling individuals. And finally,
p.valRefinedFixed and R2RefinedFixedPermutationVector with p-values and R2 of the
permutations shuffling the most important miRNAs.
}
\description{
Partial Least Squares Discriminant Analysis for isomirSeqData
}
\details{
Partial Least Squares Discriminant Analysis (PLS-DA) is a technique specifically
appropriate for analysis of high dimensionality data sets and multicollinearity
\cite{perezenciso}. PLS-DA is a supervised method (i.e. makes use of class
labels) with the aim to provide a dimension reduction strategy in a situation
where we want to relate a binary response variable (in our case young or old
status) to a set of predictor variables. Dimensionality reduction procedure is
based on orthogonal transformations of the original variables (isomiRs) into a
set of linearly uncorrelated latent variables (usually termed as components)
such that maximizes the separation between the different classes in the first
few components \cite{xia}. We used sum of squares captured by the model (R2) as
a goodness of fit measure. We implemented this method using the
\code{\link[DiscriMiner]{DiscriMiner-package}} into \code{\link{isoPLSDA}} function. The output
p-value of this function will tell about the statistical
significant of the group separation using miRNA/isomiR expression data.
}
\examples{
library(DESeq2)
data(mirData)
ids = isoCounts(mirData, iso5=TRUE, iso3=TRUE, add=TRUE, ref=TRUE)
ids = isoNorm(ids)
pls.ids = isoPLSDA(ids, "condition", nperm = 10)
cat(paste0("pval:",pls.ids$p.val))
cat(paste0("components:",pls.ids$components))
}

