% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Components plot (pairs plot)}
\usage{
isoPLSDAplot(components, groups)
}
\arguments{
  \item{components}{PLS-DA components as it comes from
  isoPLSDA main function}

  \item{groups}{vector or factor with group memberships}
}
\value{
plot
}
\description{
Components plot (pairs plot)
}

