\name{loadIso2}
\alias{loadIso2}
\title{create class}
\usage{
loadIso2(files, design, cov = 1, header = F, skip = 1)
}
\arguments{
  \item{files}{all samples}

  \item{cov}{remove sequences that have relative abundance
  lower than this number}

  \item{design}{data frame containing groups for each
  sample}

  \item{header}{files contain headers}

  \item{skip}{skip first line when reading files}
}
\description{
create class
}

