% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQNO.R
\name{LQNO}
\alias{LQNO}
\alias{dLQNO}
\alias{pLQNO}
\alias{qLQNO}
\alias{rLQNO}
\title{LQNO distribution}
\usage{
dLQNO(x, mu = 1, sigma = 1, log = FALSE)
pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rLQNO(n, mu = 1, sigma = 1)
LQNO(mu.link="log", sigma.link="log")

dLQNO(x, mu = 1, sigma = 1, log = FALSE)

pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rLQNO(n, mu = 1, sigma = 1)
}
\arguments{
\item{mu.link}{type of transformation}

\item{sigma.link}{type of transformation}

\item{x}{vector of quantiles.}

\item{mu}{vector of means.}

\item{sigma}{vector of standard deviations.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X < x) otherwise, P(X > x).}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
Linear quadratic family that assumes the following relation for the *variance*
of the normal distribution \code{Var = mu*(1+s*mu)}.
regression on mu and on the sigma (log and identity links)
}
