% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma_ex}
\alias{ma_ex}
\title{Data frame containing gene-mirna relationship}
\format{A data frame with rows sames as gene_ex_rse and
columns same as mirna_ex_rse.}
\usage{
ma_ex
}
\description{
Data frame containing gene-mirna relationship
}
\keyword{datasets}
