% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{mirna2targetscan}
\alias{mirna2targetscan}
\title{Find targets in targetscan database}
\usage{
mirna2targetscan(mirna, species = "hsa", org = NULL, keytype = NULL)
}
\arguments{
\item{mirna}{Character vector with miRNA names as in
miRBase 21.}

\item{species}{hsa or mmu supported right now.}

\item{org}{\link[AnnotationDbi:AnnotationDb]{AnnotationDbi::AnnotationDb} obejct. For example:(org.Mm.eg.db)}

\item{keytype}{Character mentioning the gene id to use.
For example, \code{ENSEMBL}.}
}
\value{
\link{data.frame} with 4 columns:
\itemize{
\item miRFamily
\item Seedmatch
\item PCT
\item entrezGene
}
}
\description{
From a list of miRNA names, find their targets
in targetscan.Hs.eg.db annotation package.
}
\examples{
library(targetscan.Hs.eg.db)
mirna2targetscan(c("hsa-miR-34c-5p"))
}
