% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirna_ex_rse}
\alias{mirna_ex_rse}
\title{Data frame containing mirna expression data}
\format{mirna expression data with 18 samples:
example of a time series data}
\usage{
mirna_ex_rse
}
\description{
Data frame containing mirna expression data
}
\keyword{datasets}
