% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirTritation}
\alias{mirTritation}
\title{Data frame containing mirna from XX paper}
\format{mirna expression data in long format.
Train and test data to use with isoCorrect}
\usage{
mirTritation
}
\description{
Data frame containing mirna from XX paper
}
\keyword{datasets}
