% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{normIso}
\alias{normIso}
\title{normalize count data}
\usage{
normIso(x, formula = ~condition)
}
\arguments{
  \item{x}{IsomirDataSeq object}

  \item{formula}{formula that will be used for DE
  library(DESeq2) data(isomiRex)
  ma<-normIso(isomiRex,formula=~condition)}
}
\description{
normalize count data
}

