% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{plotIso}
\alias{plotIso}
\title{Plot the amount of isomiRs in different samples divided by group}
\usage{
plotIso(x, type = "t5")
}
\arguments{
  \item{x}{object isomirDataSeq}

  \item{type}{string (t5,t3,add,sub) to indicate what
  isomiR change to use for the plot}
}
\value{
ggplot2 figure showing the selected isomiR changes among
samples
}
\description{
Plot the amount of isomiRs in different samples divided by
group
}
\examples{
data(isomiRexp)
plotIso(isomiRexp)
}

