% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subs.position}
\alias{subs.position}
\title{Do summary of nt substitution events}
\usage{
subs.position(table, colid)
}
\arguments{
\item{table}{object miraligner table}

\item{colid}{column to be considered}
}
\description{
Do summary of nt substitution events
}

