% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{updateIsomirDataSeq}
\alias{updateIsomirDataSeq}
\title{Update \link{IsomirDataSeq} object from version < 1.7}
\usage{
updateIsomirDataSeq(object)
}
\arguments{
\item{object}{\link{IsomirDataSeq}.}
}
\description{
In version 1.9 IsomirDataSeq object changed their
internal structure to save space and speed up
loading and downstream functions.
}
\details{
This function will update to the current structure.
}
