% Generated by roxygen2 (4.0.0): do not edit by hand
\name{collapse.mirs}
\alias{collapse.mirs}
\title{Collapse isomiRs in miRNAs}
\usage{
collapse.mirs(table, ref = F, iso5 = F, iso3 = F, add = F, mism = F,
  seed = F)
}
\arguments{
\item{table}{object miraligner table}

\item{ref}{differenciate reference miRNA from rest}

\item{iso5}{differenciate trimming at 5 miRNA from rest}

\item{iso3}{differenciate trimming at 3 miRNA from rest}

\item{add}{differenciate additions miRNA from rest}

\item{mism}{differenciate nt substitution miRNA from rest}

\item{seed}{differenciate changes in 2-7 nt from rest}
}
\description{
Collapse isomiRs in miRNAs
}

