% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,IsomirDataSeq-method}
\alias{counts<-,IsomirDataSeq,matrix-method}
\alias{normcounts}
\alias{normcounts,IsomirDataSeq-method}
\title{Accessors for the 'counts' slot of a IsomirDataSeq object.}
\usage{
\S4method{counts}{IsomirDataSeq}(object)
\S4method{normcounts}{IsomirDataSeq}(object)

\S4method{counts}{IsomirDataSeq,matrix}(object)<-value

\S4method{counts}{IsomirDataSeq,matrix}(object) <- value

\S4method{normcounts}{IsomirDataSeq}(object)

\S4method{normcounts}{IsomirDataSeq}(object) <- value
}
\arguments{
  \item{object}{a \code{IsomirDataSeq} object.}

  \item{value}{an integer matrix}
}
\description{
The counts slot holds the count data as a matrix of
non-negative integer count values, one row for each
observational unit (gene or the like), and one column for
each sample. Similar to DESeq2 object. The normalized
matrix can be obtained by \code{normcounts} method.
}
\author{
Lorena Pantano
}

