% Generated by roxygen2 (4.0.0): do not edit by hand
\name{filter.by.cov}
\alias{filter.by.cov}
\title{filter by relative abundance to reference}
\usage{
filter.by.cov(table, limit = 10)
}
\arguments{
\item{table}{object miraligner table}

\item{limit}{remove sequences lower than this number}
}
\description{
filter by relative abundance to reference
}

