\name{filter.table}
\alias{filter.table}
\title{Filter table by relative abundance to reference}
\usage{
filter.table(table, cov = 10)
}
\arguments{
  \item{table}{object miraligner table}

  \item{cov}{remove sequences that have relative abundance
  lower than this number}
}
\description{
Filter table by relative abundance to reference
}

