% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{findTargets}
\alias{findTargets}
\title{Find miRNAs target using mRNA/miRNA expression}
\usage{
findTargets(mirna_rse, gene_rse, target, summarize = "group",
  min_cor = -0.6)
}
\arguments{
\item{mirna_rse}{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} with miRNA
information. See details.}

\item{gene_rse}{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} with gene
information. See details.}

\item{target}{Matrix with miRNAs (columns) and genes (rows)
target prediction values (1 if it is a target, 0 if not).}

\item{summarize}{Character column name in colData(rse) to use to group
samples and compare betweem miRNA/gene expression.}

\item{min_cor}{Numeric cutoff for correlation value that
will be use to consider a miRNA-gene pair as valid.}
}
\value{
mirna-gene matrix
}
\description{
This function creates a matrix with rows (genes) and
columns (mirnas) with values indicating if miRNA-gene
pair is target according putative targets and negative
correlation of the expression of both molecules.
}
\examples{

data(isoExample)
mirna_ma <- matrix(rbinom(20*25, c(0, 1), 1), ncol = 20)
colnames(mirna_ma) <- rownames(mirna_ex_rse)
rownames(mirna_ma) <- rownames(gene_ex_rse)
corMat <- findTargets(mirna_ex_rse, gene_ex_rse, mirna_ma)
}
