% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoCounts}
\alias{isoCounts}
\title{create count tables from isomirs}
\usage{
isoCounts(x, ref = FALSE, iso5 = FALSE, iso3 = FALSE, add = FALSE,
  subs = FALSE, seed = FALSE, minc = 10)
}
\arguments{
\item{x}{IsomirDataSeq class}

\item{ref}{differenciate reference miRNA from rest}

\item{iso5}{differenciate trimming at 5 miRNA from rest}

\item{iso3}{differenciate trimming at 3 miRNA from rest}

\item{add}{differenciate additions miRNA from rest}

\item{subs}{differenciate nt substitution miRNA from rest}

\item{seed}{differenciate changes in 2-7 nt from rest}

\item{minc}{int minimum number of isomiR sequences}
}
\value{
count table
}
\description{
create count tables from isomirs
}
\examples{
library(DESeq2)
data(isomiRexp)
obj <- isoCounts(isomiRexp, ref=TRUE)
head(counts(obj))
}

