% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoCounts}
\alias{isoCounts}
\title{Create count matrix}
\usage{
isoCounts(x, ref = FALSE, iso5 = FALSE, iso3 = FALSE, add = FALSE,
  subs = FALSE, seed = FALSE, minc = 10)
}
\arguments{
\item{x}{IsomirDataSeq class}

\item{ref}{differenciate reference miRNA from rest}

\item{iso5}{differenciate trimming at 5 miRNA from rest}

\item{iso3}{differenciate trimming at 3 miRNA from rest}

\item{add}{differenciate additions miRNA from rest}

\item{subs}{differenciate nt substitution miRNA from rest}

\item{seed}{differenciate changes in 2-7 nt from rest}

\item{minc}{int minimum number of isomiR sequences}
}
\value{
count table
}
\description{
This function collapses isomiRs into different groups.
}
\details{
You can merge all isomiRs into miRNA by calling the function only
with the frist two parameters. You can get a table with isomiRs and
the reference miRBase sequence by calling the function with \code{ref=TRUE}.
You can get a table with 5' trimming isomiRS, miRBase reference and
the rest by calling with \code{ref=TRUE,iso5=TRUE}.
If you set up all parameters to TRUE, you will get a table for
each differnt sequence mapping to a miRNA.
}
\examples{
library(DESeq2)
data(isomiRexp)
ids <- isoCounts(isomiRexp, ref=TRUE)
head(counts(ids))
}

