% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoCounts}
\alias{isoCounts}
\title{Create count matrix with different summarizing options}
\usage{
isoCounts(ids, ref = FALSE, iso5 = FALSE, iso3 = FALSE, add = FALSE,
  subs = FALSE, seed = FALSE, minc = 1)
}
\arguments{
\item{ids}{object of class \code{\link{IsomirDataSeq}}}

\item{ref}{differentiate reference miRNA from rest}

\item{iso5}{differentiate trimming at 5 miRNA from rest}

\item{iso3}{differentiate trimming at 3 miRNA from rest}

\item{add}{differentiate additions miRNA from rest}

\item{subs}{differentiate nt substitution miRNA from rest}

\item{seed}{differentiate changes in 2-7 nt from rest}

\item{minc}{int minimum number of isomiR sequences}
}
\value{
\code{\link{IsomirDataSeq}} object with new count table.
The count matrix can be access with \code{counts(ids)}.
}
\description{
This function collapses isomiRs into different groups.
}
\details{
You can merge all isomiRs into miRNAs by calling the function only
with the first parameter \code{isoCounts(ids)}.
You can get a table with isomiRs and
the reference miRBase sequences by calling the function with \code{ref=TRUE}.
You can get a table with 5' trimming isomiRS, miRBase reference and
the rest by calling with \code{isoCounts(ids, ref=TRUE, iso5=TRUE)}.
If you set up all parameters to TRUE, you will get a table for
each different sequence mapping to a miRNA (i.e. all isomiRs).

Examples for the naming used for the isomiRs are  at
\url{http://seqcluster.readthedocs.org/mirna_annotation.html#mirna-annotation}.
}
\examples{
data(mirData)
ids <- isoCounts(mirData, ref=TRUE)
head(counts(ids))
# taking into account isomiRs at 5' end.
ids <- isoCounts(mirData, ref=TRUE, iso5=TRUE)
head(counts(ids))
}

