% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoDE}
\alias{isoDE}
\title{Do differential expression analysis with DESeq2}
\usage{
isoDE(ids, formula, ref = FALSE, iso5 = FALSE, iso3 = FALSE,
  add = FALSE, subs = FALSE, seed = FALSE)
}
\arguments{
\item{ids}{object isomiDataSeq}

\item{formula}{used for DE analysis}

\item{ref}{differenciate reference miRNA from rest}

\item{iso5}{differenciate trimming at 5 miRNA from rest}

\item{iso3}{differenciate trimming at 3 miRNA from rest}

\item{add}{differenciate additions miRNA from rest}

\item{subs}{differenciate nt substitution miRNA from rest}

\item{seed}{differenciate changes in 2-7 nt from rest}
}
\value{
\code{\link{DESeqDataSet}} object
}
\description{
This function does differential expression analysis with \code{\link{DESeq2}}.
It will return a \code{\link{DESeqResults}} object.
}
\details{
This function allows to collpase all isomiRs in different types.
Read more at \code{\link{isoCounts}}.

After collpasing isomiRs, \code{\link{DESeq2}} is used to do the differential
expression analysis. It uses the design matrix given by the user as the colData
slot to construct \code{\link{DESeqData}} object.
}
\examples{
library(DESeq2)
data(isomiRexp)
dds <- isoDE(isomiRexp, formula=~condition)
}

