% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoDE}
\alias{isoDE}
\title{Differential expression analysis with DESeq2}
\usage{
isoDE(ids, formula, ref = FALSE, iso5 = FALSE, iso3 = FALSE,
  add = FALSE, subs = FALSE, seed = FALSE)
}
\arguments{
\item{ids}{object of class \code{\link{IsomirDataSeq}}}

\item{formula}{used for DE analysis}

\item{ref}{differentiate reference miRNA from rest}

\item{iso5}{differentiate trimming at 5 miRNA from rest}

\item{iso3}{differentiate trimming at 3 miRNA from rest}

\item{add}{differentiate additions miRNA from rest}

\item{subs}{differentiate nt substitution miRNA from rest}

\item{seed}{differentiate changes in 2-7 nt from rest}
}
\value{
\code{\link[DESeq2]{DESeqDataSet}} object
}
\description{
This function does differential expression analysis with \code{\link[DESeq2]{DESeq2-package}}.
It will return a \code{\link{DESeqResults}} object.
}
\details{
This function collapses all isomiRs in different types.
Read more at \code{\link{isoCounts}}.

After that isomiRs, \code{\link[DESeq2]{DESeq2-package}} is used to do differential
expression analysis. It uses the design matrix given when the
\code{\link{IsomirDataSeq}} is created
to construct a \code{\link[DESeq2]{DESeqDataSet}} object.
}
\examples{
library(DESeq2)
data(mirData)
dds <- isoDE(mirData, formula=~condition)
}

