% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoDE}
\alias{isoDE}
\title{Differential expression analysis with DESeq2}
\usage{
isoDE(ids, formula = NULL, ...)
}
\arguments{
\item{ids}{object of class \code{\link{IsomirDataSeq}}}

\item{formula}{used for DE analysis}

\item{...}{options to pass to \link{isoCounts} including
ref, iso5, iso3, add, subs and seed parameters.}
}
\value{
\code{\link[DESeq2]{DESeqDataSet}} object.
To get the differential expression isomiRs, use \link[DESeq2]{results} from
DESeq2 package. This allows to ask for different contrast
without calling again \code{isoDE}. Read \code{results}
manual to know how to access all the information.
}
\description{
This function does differential expression analysis with
\code{\link[DESeq2]{DESeq2-package}} using the specific formula.
It will return a \code{\link[DESeq2]{DESeqDataSet}} object.
}
\details{
First, this function collapses all isomiRs in different types.
Read more at \code{\link{isoCounts}} to know the different options
available to collapse isomiRs.

After that, \code{\link[DESeq2]{DESeq2-package}} is used to do differential
expression analysis. It uses the count matrix and design experiment
stored at (\code{counts(ids)} and \code{colData(ids)})
\code{\link{IsomirDataSeq}} object
to construct a \code{\link[DESeq2]{DESeqDataSet}} object.
}
\examples{
data(mirData)
ids <- isoCounts(mirData, minc=10, mins=6)
dds <- isoDE(mirData, formula=~group)

}
