% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IsomirDataSeq-class}
\alias{IsomirDataSeq}
\alias{IsomirDataSeq-class}
\title{Class that contains all isomiRs annotation for all samples}
\description{
The \code{\link{IsomirDataSeq}} is a subclass of
\code{\link[SummarizedExperiment]{SummarizedExperiment}}
used to store the raw data, intermediate calculations and results of an
miRNA/isomiR analysis. This class stores all raw isomiRs
data for each sample, processed information,
summary for each isomiR type,
raw counts, normalized counts, and table with
experimental information for each sample.
}
\details{
\code{\link{IsomirDataSeqFromFiles}} creates this object using seqbuster output files.

Methods for this objects are \code{\link[isomiRs]{counts}} to get count matrix
and \code{\link[isomiRs]{isoSelect}}
for miRNA/isomiR selection. Functions
available for this object are \code{\link[isomiRs]{isoCounts}} for count matrix creation,
\code{\link[isomiRs]{isoNorm}} for normalization, \code{\link[isomiRs]{isoDE}} for
differential expression and \code{\link{isoPLSDA}} for clustering.
\code{\link[isomiRs]{isoPlot}} helps with basic expression plot.

\code{metadata} contains two lists: \code{rawList} is list with same
length than number of samples and stores the input files
for each file, \code{isoList} is a list with same length than
number of samples and stores information for each isomiR type summarizing
the different changes for the different isomiRs types (trimming at 3',
trimming a 5', addition and substitution).
}
\examples{
path <- system.file("extra", package="isomiRs")
fn_list <- list.files(path, full.names = TRUE)
de <- data.frame(row.names=c("f1" , "f2"), condition = c("newborn", "newborn"))
ids <- IsomirDataSeqFromFiles(fn_list, design=de)

head(counts(ids))

}

