% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IsomirDataSeq}
\alias{IsomirDataSeq}
\alias{IsomirDataSeq-class}
\title{Class that contain all isomiRs annotation for all samples}
\usage{
IsomirDataSeq(se, expList, isoList, statsList)

IsomirDataSeq(se, expList, isoList, statsList)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{expList}{list of samples with seqbuster output.}

\item{isoList}{list of samples with summarized isomiR
information for each type.}

\item{statsList}{list of samples with general isomiR information.
Could be empty list.}
}
\description{
The \code{\link{IsomirDataSeq}} is a subclass of
\code{\link[SummarizedExperiment]{SummarizedExperiment}}
used to store the raw data, intermediate calculations and results of an
miRNA/isomiR analysis.  The \code{\link{IsomirDataSeq}} class stores all raw isomiRs
data for each sample, processed information,
summary for each isomiR type,
raw counts, normalized counts, and table with
experimental information for each sample.
}
\details{
\code{\link{IsomirDataSeqFromFiles}} creates this object using seqbuster output files.

Methods for this objects are \code{\link[isomiRs]{counts}} to get count matrix
and \code{\link[isomiRs]{isoSelect}}
for miRNA/isomiR selection. Functions
available for this object are \code{\link[isomiRs]{isoCounts}} for count matrix creation,
\code{\link[isomiRs]{isoNorm}} for normalization, \code{\link[isomiRs]{isoDE}} for
differential expression and \code{\link{isoPLSDA}} for clustering.
\code{\link[isomiRs]{isoPlot}} helps with basic expression plot.
}
\examples{
\dontrun{
fn_list = c("url1", "url2")
de = data.frame(row.names=c("f1" , "f2"), condition = c("n1", "o1"))
ids <- IsomirDataSeqFromFiles(fn_list, design=de)

select(ids, "hsa-let-7a-5p")
counts(ids)[1:5, ]
}
}

