% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{IsomirDataSeqFromFiles}
\alias{IsomirDataSeqFromFiles}
\title{\code{IsomirDataSeqFromFiles} loads miRNA annotation from seqbuster tool}
\usage{
IsomirDataSeqFromFiles(files, design, rate = 0.5, header = TRUE, skip = 0,
  quiet = TRUE, ...)
}
\arguments{
\item{files}{files with the output of seqbuster tool}

\item{design}{data frame containing groups for each sample}

\item{rate}{minimum counts fraction to consider a mismatch a real mutation}

\item{header}{boolean to indicate files contain headers}

\item{skip}{skip first line when reading files}

\item{quiet}{boolean indicating to print messages
while reading files. Default \code{FALSE}.}

\item{...}{arguments provided to \code{\link[SummarizedExperiment]{SummarizedExperiment}}
including rowData.}
}
\value{
\code{\link{IsomirDataSeq}} class object.
}
\description{
This function parses
output of seqbuster tool to allow isomiRs/miRNAs analysis of samples
in different groups such as
characterization, differential expression and clustering. It creates an
\code{\link[isomiRs]{IsomirDataSeq}} object.
}
\details{
This function parses the output of \url{http://seqcluster.readthedocs.org/mirna_annotation.html}
for each sample to create a count matrix for isomiRs, miRNAs or isomiRs grouped in
types (i.e all sequences with variations at 5' but ignoring any other type). It creates
\code{\link[isomiRs]{IsomirDataSeq}} object (see link to example usage of this class)
to allow visualization, queries, differential
expression analysis and clustering.
To create the \code{\link[isomiRs]{IsomirDataSeq}}, it parses the isomiRs files, and generates
an initial matrix having all isomiRs detected among samples. As well, it creates
a summary for each isomiR type (trimming, addition and substitution) to
visualize general isomiRs distribution.
}
\examples{
path <- system.file("extra", package="isomiRs")
fn_list <- list.files(path, full.names = TRUE)
de <- data.frame(row.names=c("f1" , "f2"), condition = c("newborn", "newborn"))
ids <- IsomirDataSeqFromFiles(fn_list, design=de)

head(counts(ids))

}

