% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRexp-data.R
\docType{data}
\name{isomiRex}
\alias{isomiRex}
\title{isomirDataSeq}
\format{a \code{isomirDataSeq} class.}
\usage{
isomiRex
}
\description{
This data set is the object return by \code{loadIso}. It
mainly contains miRNA raw data from 6 samples: 3 new birth
and 3 elderly human individuals. CITE. Use \code{colData}
to see the experiment design of the data.
}
\author{
Lorena Pantano, 2015-05-19
}

