% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{isomiRex}
\alias{isomiRex}
\title{isomirDataSeq}
\format{a \code{isomirDataSeq} instance, 1 row per CpG island.}
\usage{
isomiRex
}
\description{
This data set is the object return by \code{loadIso}
}
\author{
Lorena Pantano, 2014-04-08
}

