% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRexp-data.R
\docType{data}
\name{isomiRexp}
\alias{isomiRexp}
\title{IsomirDataSeq}
\format{a \code{IsomirDataSeq} class.}
\usage{
load(isomiRex)
}
\description{
This data set is the object return by \code{IsomirDataSeqFromFiles}. It mainly
contains miRNA sequencing raw data from 6 samples: 3 newnborns and 3 elderly
human individuals (Somel et al, 2010).
Use \code{colData} to see the experiment design of the data.
}
\author{
Lorena Pantano, 2015-05-19
}

