% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{isoNetwork}
\alias{isoNetwork}
\title{Clustering miRNAs-genes pairs in similar pattern expression}
\usage{
isoNetwork(mirna_rse, gene_rse, summarize = NULL, target = NULL,
  org = NULL, enrich = NULL, genename = "ENSEMBL", min_cor = -0.6,
  min_fc = 0.5)
}
\arguments{
\item{mirna_rse}{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} with miRNA
information. See details.}

\item{gene_rse}{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} with gene
information. See details.}

\item{summarize}{Character column name in \code{colData(rse)} to use to group
samples and compare betweem miRNA/gene expression.}

\item{target}{Matrix with miRNAs (columns) and genes (rows)
target prediction (1 if it is a target, 0 if not).}

\item{org}{\link[AnnotationDbi:AnnotationDb]{AnnotationDbi::AnnotationDb} obejct. For example:(org.Mm.eg.db)}

\item{enrich}{The output of clusterProfiler of similar functions.}

\item{genename}{Character keytype of the gene
names in gene_rse object.}

\item{min_cor}{Numeric cutoff to consider a miRNA to regulate a target.}

\item{min_fc}{Numeric cutoff to consider as the minimum log2FoldChange
between groups to be considered in the analysis.}
}
\value{
list with network information
}
\description{
Clustering miRNAs-genes pairs
}
\details{
This function will correlate miRNA and gene expression data using
a specific metadata variable to group samples and detect pattern
of expression that will be annotated with GO terms.
mirna_rse and gene_rse can be created using the following code:

\code{mi_rse = SummarizedExperiment(assays=SimpleList(norm=mirna_matrix), colData, metadata=list(sign=mirna_keep))}

where, \code{mirna_matrix} is the normalized counts expression,
\code{colData} is the metadata information and \code{mirna_keep}
the list of miRNAs to be used by this function.
}
\examples{
# library(org.Mm.eg.db)
# library(clusterProfiler)
data(isoExample)
# ego <- enrichGO(row.names(assay(gene_ex_rse, "norm")),
#                 org.Mm.eg.db, "ENSEMBL", ont = "BP")
data <- isoNetwork(mirna_ex_rse, gene_ex_rse, 
                   summarize = "group", target = ma_ex,
                   enrich = ego)
isoPlotNet(data, minGenes = 5)
}
