% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoNorm}
\alias{isoNorm}
\title{Normalize count data}
\usage{
isoNorm(x, formula = ~condition)
}
\arguments{
\item{x}{IsomirDataSeq object}

\item{formula}{formula that will be used for DE}
}
\value{
IsomirSeqData object with the normalized
count matrix in a slot. The normalized matrix
can be access with counts(x, norm=TRUE).
}
\description{
This function normalizes raw count data using
rlog function from DESeq2.
}
\examples{
library(DESeq2)
data(isomiRexp)
obj <- isoNorm(isomiRexp, formula=~condition)
head(counts(obj, norm=TRUE))
}

