% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoNorm}
\alias{isoNorm}
\title{Normalize count matrix}
\usage{
isoNorm(ids, formula = ~condition)
}
\arguments{
\item{ids}{IsomirDataSeq object}

\item{formula}{formula that will be used for normalization}
}
\value{
\code{\link{IsomirDataSeq}} object with the normalized
count matrix in a slot. The normalized matrix
can be access with \code{counts(ids, norm=TRUE)}.
}
\description{
This function normalizes raw count matrix using
\code{\link[DESeq2]{rlog}} function from \code{\link[DESeq2]{DESeq2-package}}.
}
\examples{
library(DESeq2)
data(isomiRexp)
ids <- isoNorm(isomiRexp, formula=~condition)
head(counts(ids, norm=TRUE))
}

