% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoPlot}
\alias{isoPlot}
\title{Plot the amount of isomiRs in different samples}
\usage{
isoPlot(ids, type = "iso5")
}
\arguments{
\item{ids}{object isomirDataSeq}

\item{type}{string (iso5,iso3,add,subs) to indicate what isomiR
change to use for the plot}
}
\value{
ggplot2 figure showing the selected isomiR changes among samples
}
\description{
Plot the amount of isomiRs in different samples
}
\examples{
data(isomiRexp)
isoPlot(isomiRexp)
}

