% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{isoPlotNet}
\alias{isoPlotNet}
\title{Functional miRNA / gene expression profile plot}
\usage{
isoPlotNet(obj, minGenes = 2)
}
\arguments{
\item{obj}{Output from \code{\link[=isoNetwork]{isoNetwork()}}.}

\item{minGenes}{Minimum number of genes per term to be kept.}
}
\value{
Network ggplot.
}
\description{
Plot analysis from \code{\link[=isoNetwork]{isoNetwork()}}. See that function
for an example of the figure.
}
