% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{isoPlotNet}
\alias{isoPlotNet}
\title{Functional miRNA / gene expression profile plot}
\usage{
isoPlotNet(obj)
}
\arguments{
\item{obj}{output from \link{isoNetwork}}
}
\description{
Functional miRNA / gene expression profile plot
}
