% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isoPLSDA}
\alias{isoPLSDA}
\title{Arguments (included in plsDA function within DiscriMiner package)}
\usage{
isoPLSDA(variables, group, validation = NULL, learn = NULL, test = NULL,
 tol = 0.001, nperm = 400, refinment = FALSE, vip = 1.2)
}
\arguments{
\item{variables}{matrix or data frame with explanatory variables}

\item{group}{vector or factor with group memberships}

\item{validation}{type of validation, either NULL or "learntest". Default NULL}

\item{learn}{optional vector of indices for a learn-set. Only used when validation="learntest". Default NULL}

\item{test}{optional vector of indices for a test-set. Only used when validation="learntest". Default NULL}

\item{tol}{tolerance value based on maximum change of cumulative R-squared coefficient for each additional PLS component. Default tol=0.001}

\item{nperm}{number of permutations to compute the PLD-DA p-value based on R2 magnitude. Default nperm=400}

\item{refinment}{logical indicating whether a refined model, based on filtering out variables with low VIP values}

\item{vip}{Variance Importance in Projection threshole value when a refinement precess is considered. Default vip=1.2}
}
\value{
PLS model
}
\description{
Arguments (included in plsDA function within DiscriMiner package)
}

