% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Components plot (pairs plot)}
\usage{
isoPLSDAplot(components, groups)
}
\arguments{
  \item{components}{PLS-DA components as it comes from
  isoPLSDA main function}

  \item{groups}{vector or factor with group memberships}
}
\value{
plot
}
\description{
Components plot (pairs plot)
}

