% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Plot components from isoPLSDA analysis (pairs plot)}
\usage{
isoPLSDAplot(pls, n = 2)
}
\arguments{
\item{pls}{Output from \code{\link[=isoPLSDA]{isoPLSDA()}} function.}

\item{n}{Number of components to plot.}
}
\value{
\code{\link[GGally:ggpairs]{GGally::ggpairs()}} plot showing the scores
for each sample using isomiRs/miRNAs expression to explain
variation.

\link[base:data.frame]{base::data.frame} object with a first column
refering to the sample group, and the following
columns refering to the score that each sample
has for each
component from the PLS-DA analysis.
}
\description{
Plot the most significant components that come from \code{\link[=isoPLSDA]{isoPLSDA()}}
analysis together with the density of the samples scores along those components.
}
\details{
The function \code{isoPLSDAplot}
helps to visualize the results from \code{\link[=isoPLSDA]{isoPLSDA()}}.
It will plot the samples using the
significant components (t1, t2, t3 ...) from the PLS-DA analysis and the
samples score distribution along the components.
It uses \code{\link[GGally:ggpairs]{GGally::ggpairs()}}
for the plot.
}
\examples{
data(mirData)
# Only miRNAs with > 10 reads in all samples.
ids <- isoCounts(mirData, minc=10, mins=6)
ids <- isoNorm(ids, formula=~group)
pls.ids <- isoPLSDA(ids, "group", nperm = 2)
isoPLSDAplot(pls.ids)
}
