% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Plot components from isoPLSDA(pairs plot)}
\usage{
isoPLSDAplot(components, groups)
}
\arguments{
\item{components}{PLS-DA components as it comes from isoPLSDA main function}

\item{groups}{vector or factor with group memberships}
}
\value{
plot

\code{\link[ggplot2]{ggplot2-package}} object
}
\description{
Plot components from isoPLSDA(pairs plot)
}
\details{
The function \code{isoPLSDAplot}
helps to visualize the results. It will plot the samples using the
significant components (t1, t2, t3 ...) from the PLS-DA analysis and the
samples distribution along the components.
}
\examples{
data(isomiRexp)
ids = isoCounts(isomiRexp, iso5=TRUE, iso3=TRUE, add=TRUE, ref=TRUE)
ids = isoNorm(ids)
pls.ids = isoPLSDA(ids, "condition", nperm = 10)
isoPLSDAplot(pls.ids$component, colData(ids)[,"condition"])
}

