% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Plot components from isoPLSDA(pairs plot)}
\usage{
isoPLSDAplot(pls)
}
\arguments{
\item{pls}{output from \code{\link{isoPLSDA}} function.}
}
\value{
\code{\link[GGally]{ggpairs}} plot

\code{\link[ggplot2]{ggplot2-package}} object
}
\description{
Plot the most significant components that come from \code{\link{isoPLSDA}}
together with the density of the samples along those components.
}
\details{
The function \code{isoPLSDAplot}
helps to visualize the results from \code{\link{isoPLSDA}}.
It will plot the samples using the 
significant components (t1, t2, t3 ...) from the PLS-DA analysis and the 
samples distribution along the components. It uses \code{\link[GGally]{ggpairs}}
for the plot.
}
\examples{
data(mirData)
ids = isoCounts(mirData, iso5=TRUE, iso3=TRUE, add=TRUE, ref=TRUE)
ids = isoNorm(ids)
pls.ids = isoPLSDA(ids, "condition", nperm = 10)
isoPLSDAplot(pls.ids)
}

