\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Components plot (LATTICE PLOT)}
\usage{
isoPLSDAplot(components, groups)
}
\arguments{
  \item{components}{PLS-DA components as it comes from
  isoPLSDA main function}

  \item{groups}{vector or factor with group memberships}
}
\value{
plot
}
\description{
Components plot (LATTICE PLOT)
}

