% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isoPLSDAplot}
\alias{isoPLSDAplot}
\title{Components plot (LATTICE PLOT)}
\usage{
isoPLSDAplot(components, groups)
}
\arguments{
\item{components}{PLS-DA components as it comes from plsDAres main function}

\item{group}{vector or factor with group memberships}
}
\value{
plot
}
\description{
Components plot (LATTICE PLOT)
}

