% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{isoraw}
\alias{isoraw}
\alias{isoraw,IsomirDataSeq-method}
\alias{isoraw<-,IsomirDataSeq,matrix-method}
\title{Accessors for the 'isoraw' slot of a IsomirDataSeq object.}
\usage{
\S4method{isoraw}{IsomirDataSeq}(object)

\S4method{isoraw}{IsomirDataSeq,matrix}(object)<-value

\S4method{isoraw}{IsomirDataSeq}(x) <- value
}
\arguments{
  \item{value}{a list of matrix}

  \item{object}{a \code{IsomirDataSeq} object.}
}
\description{
The isoraw slot holds the raw data as a list of matrix
values from miraligner tool.
}
\author{
Lorena Pantano
}

