% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{isoSelect}
\alias{isoSelect}
\alias{isoSelect,IsomirDataSeq-method}
\alias{isoSelect.IsomirDataSeq}
\title{Method to browse an IsomirDataSeq object.}
\usage{
\S4method{isoSelect}{IsomirDataSeq}(object, mirna, norm=FALSE, minc=10)

\S4method{isoSelect}{IsomirDataSeq}(object, mirna = "", norm = FALSE,
  minc = 10)
}
\arguments{
\item{object}{a \code{IsomirDataSeq} object.}

\item{mirna}{string of the miRNA to show}

\item{norm}{TRUE return log2-normalized counts}

\item{minc}{int minimum number of isomiR reads}
}
\value{
\code{\link[S4Vectors]{DataFrame-class}}
}
\description{
This method allows to select a miRNA and all its isomiRs
from all samples.
}
\examples{
data(mirData)
isoSelect(mirData, mirna="hsa-let-7a-5p")
}
\author{
Lorena Pantano
}

