% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{isoSelect}
\alias{isoSelect}
\alias{isoSelect,IsomirDataSeq-method}
\alias{isoSelect.IsomirDataSeq}
\title{Method for browse an IsomirDataSeq object.}
\usage{
\S4method{isoSelect}{IsomirDataSeq}(object, norm=FALSE, minc=10, mirna="")

\S4method{isoSelect}{IsomirDataSeq}(object, norm = FALSE, minc = 10,
  mirna = "")
}
\arguments{
  \item{object}{a \code{IsomirDataSeq} object.}

  \item{norm}{TRUE return log2-normalized counts}

  \item{minc}{int minimum number of isomiR reads}

  \item{mirna}{string of the miRNA to show}
}
\description{
This method allows to select a miRNA and all its isomiRs
along all samples.
}
\author{
Lorena Pantano
}

