% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoTop}
\alias{isoTop}
\title{Heatmap of the top expressed isomiRs}
\usage{
isoTop(ids, top = 20)
}
\arguments{
\item{ids}{object of class \code{\link{IsomirDataSeq}}}

\item{top}{number of isomiRs/miRNAs used}
}
\description{
This function creates a heatmap with the top N
isomiRs/miRNAs. It uses the matrix under \code{counts(ids)}
and represent in a heatmap the raw counts for each sample.
}
\examples{
data(mirData)
isoTop(mirData)
}

