% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoTop}
\alias{isoTop}
\title{Plot the amount of isomiRs in different samples divided by group}
\usage{
isoTop(x, top = 20)
}
\arguments{
  \item{x}{object isomiDataSeq}

  \item{top}{number of isomiRs used}
}
\description{
Plot the amount of isomiRs in different samples divided by
group
}
\examples{
library(DESeq2)
data(isomiRexp)
isoTop(isomiRexp)
}

