% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isomiRs.R
\name{isoTop}
\alias{isoTop}
\title{Plot the top expressed isomiRs}
\usage{
isoTop(ids, top = 20)
}
\arguments{
\item{ids}{object isomiDataSeq}

\item{top}{number of isomiRs used}
}
\description{
Plot the top expressed isomiRs
}
\examples{
library(DESeq2)
data(isomiRexp)
isoTop(isomiRexp)
}

