% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadIso}
\alias{loadIso}
\title{load data}
\usage{
loadIso(files, design, cov = 1, header = F, skip = 1)
}
\arguments{
\item{files}{all samples}

\item{cov}{remove sequences that have relative abundance lower than this number}

\item{design}{data frame containing groups for each sample}

\item{header}{files contain headers}

\item{skip}{skip first line when reading files}
}
\description{
load data
}

