% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQNO.R
\name{LQNO}
\alias{LQNO}
\alias{dLQNO}
\alias{pLQNO}
\alias{qLQNO}
\alias{rLQNO}
\title{LQNO distribution}
\usage{
dLQNO(x, mu = 1, sigma = 1, log = FALSE)
pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rLQNO(n, mu = 1, sigma = 1)
LQNO(mu.link="log", sigma.link="log")

dLQNO(x, mu = 1, sigma = 1, log = FALSE)

pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rLQNO(n, mu = 1, sigma = 1)
}
\arguments{
\item{mu.link}{Type of transformation}

\item{sigma.link}{Type of transformation}

\item{x}{Vector of quantiles.}

\item{mu}{Vector of means.}

\item{sigma}{Vector of standard deviations.}

\item{log}{Logical; if TRUE, probabilities p are given as \code{log(p)}.}

\item{q}{Vector of quantiles.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are \code{P(X &lt; x)}
otherwise, \code{P(X &gt; x)}.}

\item{log.p}{Logical; if TRUE, probabilities p are given as \code{log(p)}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\value{
LQNO function
}
\description{
Linear quadratic family that assumes the following relation for the \emph{variance}
of the normal distribution \code{Var = mu*(1+s*mu)}.
regression on mu and on the sigma (log and identity links)
}
\details{
Methods adapted from:

Argyropoulos, Christos, et al. "Modeling bias and variation in
the stochastic processes of small RNA sequencing."
Nucleic Acids Research (2017).
}
\author{
Christos Argyropoulos
}
