% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirData-data.R
\docType{data}
\name{mirData}
\alias{mirData}
\title{Example of IsomirDataSeq with human brain miRNA counts data}
\format{a \code{\link{IsomirDataSeq}} class.}
\source{
Data is available from GEO dataset under accession number GSE18069. 
Samples used are: GSM450597, GSM450598, GSM450600,
GSM450604, GSM450610 and GSM450609 .

Every sample was analyzed with seqbuster tool, see 
\url{http://seqcluster.readthedocs.org/mirna_annotation.html} for
more details. You can get same files running the small RNA-seq
pipeline from \url{https://github.com/chapmanb/bcbio-nextgen}.

Adapter removal with the following parameters:
\code{adrec fastq_file TCGTATGCCGTCTT 8 1 0.34}

miRNAs annotation with the following parameters:
\code{miraligner adrec_output mirbase_files hsa 1 3 3 out_prefix}

The data was created with \code{isomiRs-package} package:

\code{library(isomiRs)}

\code{fns <- c("GSM450597.mirna", "GSM450598.mirna", "GSM450600.mirna",
"GSM450604.mirna", "GSM450610.mirna", "GSM450609.mirna")}

\code{design <- data.frame(condition=c('nb', 'nb', 'nb', 
'o', 'o', 'o'))}

\code{mirData <- IsomirDataSeqFromFiles(fns, design)}
}
\usage{
data("mirData")
}
\description{
This data set is the object return by \code{\link{IsomirDataSeqFromFiles}}.
It contains miRNA count data from 6 samples: 3 newborns and 3 elderly
human individuals (Somel et al, 2010). 
Use \code{colData} to see the experiment design.
}
\author{
Lorena Pantano, 2015-05-19
}
\references{
Mehmet Somel et al. MicroRNA, mRNA, and protein expression link 
development and aging in human and macaque brain.
\code{Genome Research,20(9):1207-1218,2010.doi:10.1101/gr.106849.110}
}

