% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{mirna2targetscan}
\alias{mirna2targetscan}
\title{Find targets in targetscan database}
\usage{
mirna2targetscan(mirna, org = NULL, keytype = NULL)
}
\arguments{
\item{mirna}{Character vector with miRNA names as in
miRBase 21.}
}
\value{
\link{data.frame} with 4 columns:
\itemize{
\item miRFamily
\item Seedmatch
\item PCT
\item entrezGene
}
}
\description{
From a list of miRNA names, find their targets
in \link{targetscan.Hs.eg.db} annotation package.
}
\examples{
mirna2targetscan(c("hsa-miR-34c-5p"))
}
