% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mirTritation}
\alias{mirTritation}
\title{Data frame containing mirna from Argyropoulos's paper}
\format{mirna expression data in long format.
Train and test data to use with isoCorrect}
\usage{
mirTritation
}
\description{
Argyropoulos, Christos, et al. "Modeling bias and variation in
the stochastic processes of small RNA sequencing."
Nucleic Acids Research (2017).
}
\keyword{datasets}
