\name{plotIso}
\alias{plotIso}
\title{Plot the amount of isomiRs in different samples divided by group}
\usage{
plotIso(x, type = "t5")
}
\arguments{
  \item{x}{object isomiDataSeq}

  \item{type}{string (t5,t3,add,sub) to indicate what
  change to see}
}
\description{
Plot the amount of isomiRs in different samples divided by
group
}

