\name{plotTop}
\alias{plotTop}
\title{Plot the amount of isomiRs in different samples divided by group}
\usage{
plotTop(x, top = 20)
}
\arguments{
  \item{x}{object isomiDataSeq}

  \item{top}{number of isomiRs used}
}
\description{
Plot the amount of isomiRs in different samples divided by
group
}

